Refactor: 사용자 인증 흐름 리팩토링 및 상태 관리 개선
- Supabase 클라이언트 직접 호출을 중앙 집중형 Zustand 스토어로 대체해 불필요한 서버 요청 방지
- ClientProvider 컴포넌트 도입
