+++ src/app/auth/callback/page.tsx
@@ -1,7 +1,6 @@
 "use client";
 // pages/auth/callback.tsx
 import { useEffect } from "react";
-import { createClient } from "@/shared/utils/supabase/client";
 import { useRouter } from "next/navigation";
 import { SignIn } from "@/features/sign-in/model/signIn.service";
 import { toast } from "react-toastify";
@@ -9,32 +8,16 @@ import { useAuthStore } from "@/shared/stores/useUserStore";
 
 export default function AuthCallback() {
   const router = useRouter();
-  const supabase = createClient();
-  const setUser = useAuthStore((state) => state.setUser);
+  const user = useAuthStore((state) => state.user);
 
   useEffect(() => {
     const handleAuth = async () => {
-      const { data, error } = await supabase.auth.getUser();
-      if (error) {
-        console.error(error);
-        return;
-      }
-
-      if (data?.user) {
-        const requestData = {
-          id: data.user.id,
-          email: data.user.email,
-          name: data.user.user_metadata.name,
-          profileImage: data.user.user_metadata.avatar_url || null,
-        };
-
-        const response = await SignIn(requestData);
+      if (user) {
+        const response = await SignIn(user);
 
         if (response?.message === "success") {
           toast.success("로그인되었습니다.");
           router.replace(`${process.env.NEXT_PUBLIC_BASE_URL}`);
-          const userData = response.data;
-          setUser(userData);
         }
 
         if (response?.error) {
@@ -44,7 +27,7 @@ export default function AuthCallback() {
     };
 
     handleAuth();
-  }, [router]);
+  }, [user, router]);
 
   return <p>로그인 중입니다...</p>;
 }
