+++ src/app/providers/ClientProvider.tsx
@@ -1,0 +1,18 @@
+"use client";
+
+import { useAuthStore } from "@/shared/stores/useUserStore";
+import { useEffect } from "react";
+
+export default function ClientProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const fetchUser = useAuthStore((state) => state.fetchUser);
+
+  useEffect(() => {
+    fetchUser();
+  }, [fetchUser]);
+
+  return <>{children}</>;
+}
